requirements.txt
# --- Core Deep Learning ---
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0

# --- Vector Database ---
faiss-cpu>=1.7.4

# --- Audio & Image Processing ---
librosa>=0.10.0
Pillow>=9.5.0
torchcodec>=0.1.0  # Crucial for the audio decoding we fixed for CLAP

# --- Data & Dataset Management ---
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
datasets>=2.12.0   # Required for Hugging Face AudioCaps loading
tqdm>=4.65.0

# --- API (From Friend's Branch) ---
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
flask>=3.0.0

# --- Demo UI & Audio Export ---
soundfile>=0.12.0

# --- Environment & Testing ---
pytest>=7.3.1
httpx>=0.24.0
python-dotenv>=1.0.0